AMOYRM2/requirements.txt
python-telegram-bot==22.1
httpx==0.28.1  
python-dotenv==1.1.1  